+++ src/ui/activities/filetransfer/components/popups/delete.rs
@@ -1,0 +1,72 @@
+use tui_realm_stdlib::Radio;
+use tuirealm::command::{Cmd, CmdResult, Direction};
+use tuirealm::event::{Key, KeyEvent, KeyModifiers};
+use tuirealm::props::{Alignment, BorderType, Borders, Color};
+use tuirealm::{Component, Event, MockComponent, NoUserEvent, State, StateValue};
+
+use crate::ui::activities::filetransfer::{Msg, TransferMsg, UiMsg};
+
+#[derive(MockComponent)]
+pub struct DeletePopup {
+    component: Radio,
+}
+
+impl DeletePopup {
+    pub fn new(color: Color) -> Self {
+        Self {
+            component: Radio::default()
+                .borders(
+                    Borders::default()
+                        .color(color)
+                        .modifiers(BorderType::Rounded),
+                )
+                .foreground(color)
+                .choices(["Yes", "No"])
+                .value(1)
+                .title("Delete file(s)?", Alignment::Center),
+        }
+    }
+}
+
+impl Component<Msg, NoUserEvent> for DeletePopup {
+    fn on(&mut self, ev: Event<NoUserEvent>) -> Option<Msg> {
+        match ev {
+            Event::Keyboard(KeyEvent {
+                code: Key::Left, ..
+            }) => {
+                self.perform(Cmd::Move(Direction::Left));
+                Some(Msg::None)
+            }
+            Event::Keyboard(KeyEvent {
+                code: Key::Right, ..
+            }) => {
+                self.perform(Cmd::Move(Direction::Right));
+                Some(Msg::None)
+            }
+            Event::Keyboard(KeyEvent { code: Key::Esc, .. }) => {
+                Some(Msg::Ui(UiMsg::CloseDeletePopup))
+            }
+            Event::Keyboard(KeyEvent {
+                code: Key::Char('y'),
+                modifiers: KeyModifiers::NONE,
+            }) => Some(Msg::Transfer(TransferMsg::DeleteFile)),
+            Event::Keyboard(KeyEvent {
+                code: Key::Char('n'),
+                modifiers: KeyModifiers::NONE,
+            }) => Some(Msg::Ui(UiMsg::CloseDeletePopup)),
+            Event::Keyboard(KeyEvent {
+                code: Key::Enter, ..
+            }) => {
+                if matches!(
+                    self.perform(Cmd::Submit),
+                    CmdResult::Submit(State::One(StateValue::Usize(0)))
+                ) {
+                    Some(Msg::Transfer(TransferMsg::DeleteFile))
+                } else {
+                    Some(Msg::Ui(UiMsg::CloseDeletePopup))
+                }
+            }
+            _ => None,
+        }
+    }
+}
