+++ src/ui/activities/filetransfer/components/popups/error.rs
@@ -1,0 +1,74 @@
+use tui_realm_stdlib::Paragraph;
+use tuirealm::event::{Key, KeyEvent};
+use tuirealm::props::{Alignment, BorderType, Borders, Color, TextSpan};
+use tuirealm::{Component, Event, MockComponent, NoUserEvent};
+
+use crate::ui::activities::filetransfer::{Msg, UiMsg};
+
+#[derive(MockComponent)]
+pub struct ErrorPopup {
+    component: Paragraph,
+}
+
+impl ErrorPopup {
+    pub fn new<S: AsRef<str>>(text: S, color: Color) -> Self {
+        Self {
+            component: Paragraph::default()
+                .alignment(Alignment::Center)
+                .borders(
+                    Borders::default()
+                        .color(color)
+                        .modifiers(BorderType::Rounded),
+                )
+                .foreground(color)
+                .text([TextSpan::from(text.as_ref())])
+                .wrap(true),
+        }
+    }
+}
+
+impl Component<Msg, NoUserEvent> for ErrorPopup {
+    fn on(&mut self, ev: Event<NoUserEvent>) -> Option<Msg> {
+        match ev {
+            Event::Keyboard(KeyEvent {
+                code: Key::Esc | Key::Enter,
+                ..
+            }) => Some(Msg::Ui(UiMsg::CloseErrorPopup)),
+            _ => None,
+        }
+    }
+}
+
+#[derive(MockComponent)]
+pub struct FatalPopup {
+    component: Paragraph,
+}
+
+impl FatalPopup {
+    pub fn new<S: AsRef<str>>(text: S, color: Color) -> Self {
+        Self {
+            component: Paragraph::default()
+                .alignment(Alignment::Center)
+                .borders(
+                    Borders::default()
+                        .color(color)
+                        .modifiers(BorderType::Rounded),
+                )
+                .foreground(color)
+                .text([TextSpan::from(text.as_ref())])
+                .wrap(true),
+        }
+    }
+}
+
+impl Component<Msg, NoUserEvent> for FatalPopup {
+    fn on(&mut self, ev: Event<NoUserEvent>) -> Option<Msg> {
+        match ev {
+            Event::Keyboard(KeyEvent {
+                code: Key::Esc | Key::Enter,
+                ..
+            }) => Some(Msg::Ui(UiMsg::CloseFatalPopup)),
+            _ => None,
+        }
+    }
+}
